requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.2
pydantic>=2.10.0
pydantic_ai==0.2.9
python-dotenv==1.0.0
httpx>=0.27.0,<0.28.0
openai>=1.75.0
google-generativeai==0.2.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.0.0
pytest==7.4.2
requests>=2.32.2
